rename files and folders, delete commented code
