+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from './components/header';
-import Main from './components/main';
-import Footer from './components/footer/';
+import Header from './components/Header';
+import Main from './components/Main';
+import Footer from './components/Footer';
 import './App.css';
 
 function App() {
