+++ src/components/Banner/styles.css
@@ -1,6 +1,5 @@
 .banner-container {
   position: relative;
-  /* opacity: 0.7; */
   height: 300px;
   background-image: url('../../img/run_zykova_2.jpg');
   background-size: 100%;
