+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import UserProvider from './components/auth/UserProvider';
 import { BrowserRouter } from 'react-router-dom';
+import UserProvider from './components/Signup/UserProvider';
 import App from './App';
+import './index.css';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
